Fix - Delete unused code; - Fix calloc using; - Move code to a new file; - Fix test issues.
