+++ libs/compiler/AST.h
@@ -651,34 +651,6 @@ size_t expression_inline_get_size(const node *const nd);
  */
 node expression_inline_get_substmt(const node *const nd, const size_t index);
 
-/**
- *	Get arguments amount of inline expression
- *
- *	@param	nd				inline expression
- *
- *	@return	Arguments amount
- */
-size_t expression_inline_get_arguments_amount(const node *const nd);
-
-/**
- *	Get argument of inline expression by index
- *
- *	@param	nd				inline expression
- *	@param	index			Argument index
- *
- *	@return	Argument
- */
-node expression_inline_get_argument(const node *const nd, const size_t index);
-
-/**
- *	Get called expression of inline expression
- *
- *	@param	nd				Inline expression
- *
- *	@return	Called expression
- */
-node expression_inline_get_callee(const node *const nd);
-
 
 /**
  *	Get statement class
